+++ src/filetransfer/scp_transfer.rs
@@ -318,7 +318,7 @@ impl FileTransfer for ScpFileTransfer {
                 Err(err) => {
                     return Err(FileTransferError::new_ex(
                         FileTransferErrorType::BadAddress,
-                        format!("{}", err),
+                        err.to_string(),
                     ))
                 }
             };
@@ -349,7 +349,7 @@ impl FileTransfer for ScpFileTransfer {
             Err(err) => {
                 return Err(FileTransferError::new_ex(
                     FileTransferErrorType::ConnectionError,
-                    format!("{}", err),
+                    err.to_string(),
                 ))
             }
         };
@@ -359,7 +359,7 @@ impl FileTransfer for ScpFileTransfer {
         if let Err(err) = session.handshake() {
             return Err(FileTransferError::new_ex(
                 FileTransferErrorType::ConnectionError,
-                format!("{}", err),
+                err.to_string(),
             ));
         }
         let username: String = match username {
@@ -381,7 +381,7 @@ impl FileTransfer for ScpFileTransfer {
                 ) {
                     return Err(FileTransferError::new_ex(
                         FileTransferErrorType::AuthenticationFailed,
-                        format!("{}", err),
+                        err.to_string(),
                     ));
                 }
             }
@@ -393,7 +393,7 @@ impl FileTransfer for ScpFileTransfer {
                 ) {
                     return Err(FileTransferError::new_ex(
                         FileTransferErrorType::AuthenticationFailed,
-                        format!("{}", err),
+                        err.to_string(),
                     ));
                 }
             }
@@ -425,7 +425,7 @@ impl FileTransfer for ScpFileTransfer {
                     }
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ConnectionError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -495,7 +495,7 @@ impl FileTransfer for ScpFileTransfer {
                     }
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -537,7 +537,7 @@ impl FileTransfer for ScpFileTransfer {
                     }
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -576,7 +576,7 @@ impl FileTransfer for ScpFileTransfer {
                     }
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -613,7 +613,7 @@ impl FileTransfer for ScpFileTransfer {
                     }
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -650,7 +650,7 @@ impl FileTransfer for ScpFileTransfer {
                     }
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -692,7 +692,7 @@ impl FileTransfer for ScpFileTransfer {
                     }
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -743,7 +743,7 @@ impl FileTransfer for ScpFileTransfer {
                     }
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -764,7 +764,7 @@ impl FileTransfer for ScpFileTransfer {
                     Ok(output) => Ok(output),
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -823,7 +823,7 @@ impl FileTransfer for ScpFileTransfer {
                     Ok(channel) => Ok(Box::new(BufWriter::with_capacity(65536, channel))),
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }
@@ -846,7 +846,7 @@ impl FileTransfer for ScpFileTransfer {
                     Ok(reader) => Ok(Box::new(BufReader::with_capacity(65536, reader.0))),
                     Err(err) => Err(FileTransferError::new_ex(
                         FileTransferErrorType::ProtocolError,
-                        format!("{}", err),
+                        err.to_string(),
                     )),
                 }
             }